moved load_data_files to a helper files to be used globally
